Basic comments resolved
