+++ requirements.txt
@@ -7,8 +7,7 @@ numpy >= 1.7
 pandas >= 0.20.3
 param >=1.5.0,<2.0
 pillow
+pyspark[pyspark] >= 2.0
 scikit-image
 toolz
 xarray >= 0.9.6
-
-pyspark>=2.0
